move store (and mock) into own files
