+++ service/history/workflow/update/registry_test.go
@@ -36,55 +36,13 @@ import (
 	protocolpb "go.temporal.io/api/protocol/v1"
 	"go.temporal.io/api/serviceerror"
 	updatepb "go.temporal.io/api/update/v1"
+
 	"go.temporal.io/server/service/history/consts"
 
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/internal/effect"
 	"go.temporal.io/server/service/history/workflow/update"
 )
-
-type mockUpdateStore struct {
-	update.Store
-	VisitUpdatesFunc               func(visitor func(updID string, updInfo *persistencespb.UpdateInfo))
-	GetUpdateOutcomeFunc           func(context.Context, string) (*updatepb.Outcome, error)
-	GetCurrentVersionFunc          func() int64
-	IsWorkflowExecutionRunningFunc func() bool
-}
-
-func (m mockUpdateStore) VisitUpdates(
-	visitor func(updID string, updInfo *persistencespb.UpdateInfo),
-) {
-	m.VisitUpdatesFunc(visitor)
-}
-
-func (m mockUpdateStore) GetUpdateOutcome(
-	ctx context.Context,
-	updateID string,
-) (*updatepb.Outcome, error) {
-	return m.GetUpdateOutcomeFunc(ctx, updateID)
-}
-
-func (m mockUpdateStore) GetCurrentVersion() int64 {
-	if m.GetCurrentVersionFunc == nil {
-		return 0
-	}
-	return m.GetCurrentVersionFunc()
-}
-
-func (m mockUpdateStore) IsWorkflowExecutionRunning() bool {
-	if m.IsWorkflowExecutionRunningFunc == nil {
-		return true
-	}
-	return m.IsWorkflowExecutionRunningFunc()
-}
-
-var emptyUpdateStore = mockUpdateStore{
-	VisitUpdatesFunc: func(func(updID string, updInfo *persistencespb.UpdateInfo)) {
-	},
-	GetUpdateOutcomeFunc: func(context.Context, string) (*updatepb.Outcome, error) {
-		return nil, serviceerror.NewNotFound("not found")
-	},
-}
 
 func TestFind(t *testing.T) {
 	t.Parallel()
